+++ src/pynwb/icephys.py
@@ -79,25 +79,63 @@ class PatchClampSeries(TimeSeries):
     (this class should not be instantiated directly).
     '''
 
-    __nwbfields__ = ('electrode',
-                     'gain',
-                     'stimulus_description',
-                     'sweep_number')
+    __nwbfields__ = (
+        'electrode',
+        'gain',
+        'stimulus_description',
+        'sweep_number',
+    )
 
-    @docval(*get_docval(TimeSeries.__init__, 'name'),  # required
-            {'name': 'data', 'type': ('array_data', 'data', TimeSeries),  # required
-             'doc': 'The data values. The first dimension must be time.'},
-            {'name': 'unit', 'type': str, 'doc': 'The base unit of measurement (should be SI unit)'},  # required
-            {'name': 'electrode', 'type': IntracellularElectrode,  # required
-             'doc': 'IntracellularElectrode group that describes the electrode that was used to apply '
-                     'or record this data.'},
-            {'name': 'gain', 'type': float, 'doc': 'Units: Volt/Amp (v-clamp) or Volt/Volt (c-clamp)'},  # required
-            {'name': 'stimulus_description', 'type': str, 'doc': 'the stimulus name/protocol', 'default': "N/A"},
-            *get_docval(TimeSeries.__init__, 'resolution', 'conversion', 'timestamps', 'starting_time', 'rate',
-                        'comments', 'description', 'control', 'control_description', 'offset'),
-            {'name': 'sweep_number', 'type': (int, 'uint32', 'uint64'),
-             'doc': 'Sweep number, allows for grouping different PatchClampSeries together '
-                    'via the sweep_table', 'default': None})
+    @docval(
+        *get_docval(TimeSeries.__init__, 'name'),  # required
+        {
+            "name": "data",
+            "type": ("array_data", "data", TimeSeries),
+            "doc": "The data values. The first dimension must be time.",
+            "shape": (None,),
+        },   # required
+        {
+            'name': 'unit',
+            'type': str,
+            'doc': 'The base unit of measurement (should be SI unit)',
+        },  # required
+        {
+            'name': 'electrode',
+            'type': IntracellularElectrode,
+            'doc': 'IntracellularElectrode group that describes the electrode that was '
+                   'used to apply or record this data.',
+        },    # required
+        {
+            'name': 'gain',
+            'type': float,
+            'doc': 'Units: Volt/Amp (v-clamp) or Volt/Volt (c-clamp)',
+        },  # required
+        {
+            'name': 'stimulus_description',
+            'type': str,
+            'doc': 'the stimulus name/protocol',
+            'default': "N/A",
+        },
+        *get_docval(
+            TimeSeries.__init__,
+            'resolution',
+            'conversion',
+            'timestamps',
+            'starting_time',
+            'rate',
+            'comments',
+            'description',
+            'control',
+            'control_description',
+            'offset',
+        ),
+        {
+            'name': 'sweep_number', 'type': (int, 'uint32', 'uint64'),
+            'doc': 'Sweep number, allows for grouping different PatchClampSeries '
+                   'together via the sweep_table',
+            'default': None,
+        }
+    )
     def __init__(self, **kwargs):
         name, data, unit, stimulus_description = popargs('name', 'data', 'unit', 'stimulus_description', kwargs)
         electrode, gain, sweep_number = popargs('electrode', 'gain', 'sweep_number', kwargs)
@@ -468,26 +506,33 @@ class IntracellularRecordingsTable(AlignedDynamicTable):
 
         super().__init__(**kwargs)
 
-    @docval({'name': 'electrode', 'type': IntracellularElectrode, 'doc': 'The intracellular electrode used'},
-            {'name': 'stimulus_start_index', 'type': int, 'doc': 'Start index of the stimulus', 'default': None},
-            {'name': 'stimulus_index_count', 'type': int, 'doc': 'Stop index of the stimulus', 'default': None},
-            {'name': 'stimulus', 'type': TimeSeries,
-             'doc': 'The TimeSeries (usually a PatchClampSeries) with the stimulus',
-             'default': None},
-            {'name': 'response_start_index', 'type': int, 'doc': 'Start index of the response', 'default': None},
-            {'name': 'response_index_count', 'type': int, 'doc': 'Stop index of the response', 'default': None},
-            {'name': 'response', 'type': TimeSeries,
-             'doc': 'The TimeSeries (usually a PatchClampSeries) with the response',
-             'default': None},
-            {'name': 'electrode_metadata', 'type': dict,
-             'doc': 'Additional electrode metadata to be stored in the electrodes table', 'default': None},
-            {'name': 'stimulus_metadata', 'type': dict,
-             'doc': 'Additional stimulus metadata to be stored in the stimuli table', 'default': None},
-            {'name': 'response_metadata', 'type': dict,
-             'doc': 'Additional resposnse metadata to be stored in the responses table', 'default': None},
-            returns='Integer index of the row that was added to this table',
-            rtype=int,
-            allow_extra=True)
+    @docval(
+        {
+            "name": "electrode",
+            "type": IntracellularElectrode,
+            "doc": "The intracellular electrode used",
+            "default": None,
+        },
+        {'name': 'stimulus_start_index', 'type': int, 'doc': 'Start index of the stimulus', 'default': None},
+        {'name': 'stimulus_index_count', 'type': int, 'doc': 'Stop index of the stimulus', 'default': None},
+        {'name': 'stimulus', 'type': TimeSeries,
+         'doc': 'The TimeSeries (usually a PatchClampSeries) with the stimulus',
+         'default': None},
+        {'name': 'response_start_index', 'type': int, 'doc': 'Start index of the response', 'default': None},
+        {'name': 'response_index_count', 'type': int, 'doc': 'Stop index of the response', 'default': None},
+        {'name': 'response', 'type': TimeSeries,
+         'doc': 'The TimeSeries (usually a PatchClampSeries) with the response',
+         'default': None},
+        {'name': 'electrode_metadata', 'type': dict,
+         'doc': 'Additional electrode metadata to be stored in the electrodes table', 'default': None},
+        {'name': 'stimulus_metadata', 'type': dict,
+         'doc': 'Additional stimulus metadata to be stored in the stimuli table', 'default': None},
+        {'name': 'response_metadata', 'type': dict,
+         'doc': 'Additional resposnse metadata to be stored in the responses table', 'default': None},
+        returns='Integer index of the row that was added to this table',
+        rtype=int,
+        allow_extra=True,
+    )
     def add_recording(self, **kwargs):
         """
         Add a single recording to the IntracellularRecordingsTable table.
@@ -508,6 +553,14 @@ class IntracellularRecordingsTable(AlignedDynamicTable):
                                                                        'response',
                                                                        kwargs)
         electrode = popargs('electrode', kwargs)
+
+        # if electrode is not provided, take from stimulus or response object
+        if electrode is None:
+            if stimulus:
+                electrode = stimulus.electrode
+            elif response:
+                electrode = response.electrode
+
         # Confirm that we have at least a valid stimulus or response
         if stimulus is None and response is None:
             raise ValueError("stimulus and response cannot both be None.")
@@ -542,8 +595,10 @@ class IntracellularRecordingsTable(AlignedDynamicTable):
             #     warnings.warn("sweep_number are usually expected to be the same for PatchClampSeries type "
             #                   "stimulus and response pairs in an intracellular recording.")
             if response.electrode != stimulus.electrode:
-                raise ValueError("electrodes are usually expected to be the same for PatchClampSeries type "
-                                 "stimulus and response pairs in an intracellular recording.")
+                raise ValueError(
+                    "electrodes are usually expected to be the same for PatchClampSeries type stimulus and response "
+                    "pairs in an intracellular recording."
+                )
 
         # Compile the electrodes table data
         electrodes = copy(popargs('electrode_metadata', kwargs))